从数据库读取电影类别信息形成列表，添加电影页面框架

+++ src/main/webapp/Movies.jsp
@@ -30,7 +30,7 @@
                     <td>${name}</td>
                     <td>${info}</td>
                     <td>${type}</td>
-                    <td><a href="movies-delete?id=${id}">删除</a></td>
+                    <td><a href="movies-delete?id=${id}" onClick="return confirm('确定删除?');">删除</a></td>
                 </tr>
             </s:iterator>
         </table>
